refactor(diff): DiffView 拆分职责，收敛行内评论渲染/zone 装配/业务 hooks
DiffView.tsx 2647 → 380 行，退化为组合根（调 hooks + 渲染布局），行为零变化。

- zones/mountInlineZones：抽象评论 zone 与草稿 zone 共用的通用 view-zone 装配
  （双层 dom/inner、stopPropagation 接管、宽度/横滚/高度同步、removeZone/unmount 清理）
- inline-comments/InlineCommentZone：行内评论渲染独立成域
- blame/ · DiffPane · DiffScopeSelect · DiffStatus · DraftZoneList · diff-types 各自拆出
- hooks/（+ index barrel）：数据流与三段 zone 装配 effect 拆成 12 个聚焦 hook
- tabs/shared/：CommentItem(tab) 与 InlineCommentZone(diff) 共用 useCommentThread
  （reply/edit/delete 状态机）与 CommentMarkdown（markdown 样板）
- i18n：评论操作 key 合并进 commentsPanel.*，四 locale 各删 9 个 diffView.* 重复项
