+++ apps/desktop/src/renderer/src/i18n/locales/de-DE.json
@@ -271,19 +271,11 @@
     "blameChangeRangeTitle": "Dieser Bereich wurde durch diesen PR geändert",
     "blameFailed": "blame fehlgeschlagen",
     "blameFailedNamed": "blame für {{path}} fehlgeschlagen",
-    "deleteCommentModalMessage": "Dadurch wird dieser Kommentar im Remote gelöscht und kann nicht rückgängig gemacht werden. Fortfahren?",
-    "deleteCommentModalTitle": "Kommentar löschen",
-    "deleteCommentTitle": "Eigenen veröffentlichten Kommentar löschen (mit Remote synchronisiert)",
-    "deleteFailed": "Löschen fehlgeschlagen: {{error}}",
-    "deleting": "Löschen…",
     "diffRenderFailed": "Diff-Darstellung fehlgeschlagen: {{message}}",
     "diffRenderFailedHint": "Ein Dateiwechsel oder erneuter Versuch behebt das Problem meist. Der zugrunde liegende Fehler wurde in der console protokolliert.",
-    "dismissErrorAria": "Fehler schließen",
     "dismissNotificationTitle": "Diese Benachrichtigung schließen",
-    "editCommentTitle": "Eigenen veröffentlichten Kommentar bearbeiten (mit Remote synchronisiert)",
     "fileCount_one": "{{count}} Datei",
     "fileCount_other": "{{count}} Dateien",
-    "gotIt": "Verstanden",
     "loadChangedFilesFailed": "Geänderte Dateien konnten nicht geladen werden",
     "loadCommentsFailed": "Kommentare konnten nicht geladen werden",
     "loadingContentHint": "git blob wird aus dem lokalen Mirror gelesen; große oder binäre Dateien können langsamer sein",
@@ -294,7 +286,6 @@
     "preparingSync": "Synchronisierung wird vorbereitet",
     "readFileContentFailed": "Dateiinhalt konnte nicht gelesen werden",
     "readFileContentFailedNamed": "Inhalt von {{path}} konnte nicht gelesen werden",
-    "reply": "Antworten",
     "resizeFileListTitle": "Ziehen, um die Breite des Dateibaums anzupassen",
     "retry": "Erneut versuchen",
     "scopeAll": "Alle Änderungen",
