+++ apps/desktop/src/renderer/src/i18n/locales/en-US.json
@@ -271,19 +271,11 @@
     "blameChangeRangeTitle": "This range was changed by this PR",
     "blameFailed": "Blame failed",
     "blameFailedNamed": "Blame failed for {{path}}",
-    "deleteCommentModalMessage": "This will delete this comment on the remote and cannot be undone. Continue?",
-    "deleteCommentModalTitle": "Delete comment",
-    "deleteCommentTitle": "Delete your own published comment (synced remotely)",
-    "deleteFailed": "Delete failed: {{error}}",
-    "deleting": "Deleting…",
     "diffRenderFailed": "Diff render failed: {{message}}",
     "diffRenderFailedHint": "Switching files or retrying usually recovers. The underlying error has been logged to the console.",
-    "dismissErrorAria": "Dismiss error",
     "dismissNotificationTitle": "Dismiss this notification",
-    "editCommentTitle": "Edit your own published comment (synced remotely)",
     "fileCount_one": "{{count}} file",
     "fileCount_other": "{{count}} files",
-    "gotIt": "Got it",
     "loadChangedFilesFailed": "Failed to load changed files",
     "loadCommentsFailed": "Failed to load comments",
     "loadingContentHint": "Reading git blob from the local mirror; large or binary files may be slower",
@@ -294,7 +286,6 @@
     "preparingSync": "Preparing sync",
     "readFileContentFailed": "Failed to read file content",
     "readFileContentFailedNamed": "Failed to read content of {{path}}",
-    "reply": "Reply",
     "resizeFileListTitle": "Drag to resize file tree",
     "retry": "Retry",
     "scopeAll": "All changes",
