+++ apps/desktop/src/renderer/src/i18n/locales/ja-JP.json
@@ -264,18 +264,10 @@
     "blameChangeRangeTitle": "この範囲はこの PR で変更されました",
     "blameFailed": "blame に失敗しました",
     "blameFailedNamed": "{{path}} の blame に失敗しました",
-    "deleteCommentModalMessage": "この操作はリモートのこのコメントを削除し、元に戻せません。続行しますか？",
-    "deleteCommentModalTitle": "コメントを削除",
-    "deleteCommentTitle": "自分が公開したコメントを削除（リモートと同期）",
-    "deleteFailed": "削除に失敗しました: {{error}}",
-    "deleting": "削除中…",
     "diffRenderFailed": "差分の表示に失敗しました: {{message}}",
     "diffRenderFailedHint": "ファイルを切り替えるか再試行すると通常は復旧します。元のエラーは console に記録されています。",
-    "dismissErrorAria": "エラーを閉じる",
     "dismissNotificationTitle": "この通知を閉じる",
-    "editCommentTitle": "自分が公開したコメントを編集（リモートと同期）",
     "fileCount_other": "{{count}} ファイル",
-    "gotIt": "了解",
     "loadChangedFilesFailed": "変更ファイルの読み込みに失敗しました",
     "loadCommentsFailed": "コメントの読み込みに失敗しました",
     "loadingContentHint": "ローカルミラーから git blob を読み込んでいます。大きいファイルやバイナリファイルは時間がかかる場合があります",
@@ -286,7 +278,6 @@
     "preparingSync": "同期を準備中",
     "readFileContentFailed": "ファイル内容の読み込みに失敗しました",
     "readFileContentFailedNamed": "{{path}} の内容の読み込みに失敗しました",
-    "reply": "返信",
     "resizeFileListTitle": "ドラッグしてファイルツリーの幅を調整",
     "retry": "再試行",
     "scopeAll": "すべての変更",
