+++ apps/desktop/src/renderer/src/i18n/locales/zh-CN.json
@@ -264,18 +264,10 @@
     "blameChangeRangeTitle": "此区段为本 PR 引入的改动",
     "blameFailed": "blame 失败",
     "blameFailedNamed": "{{path}} blame 失败",
-    "deleteCommentModalMessage": "此操作会删除远端的这条评论，且无法恢复。确定继续吗？",
-    "deleteCommentModalTitle": "删除评论",
-    "deleteCommentTitle": "删除自己发布的评论 (远端同步)",
-    "deleteFailed": "删除失败：{{error}}",
-    "deleting": "删除中…",
     "diffRenderFailed": "diff 渲染失败：{{message}}",
     "diffRenderFailedHint": "切换文件 / 重试通常能恢复。底层异常已记录到 console。",
-    "dismissErrorAria": "关闭错误",
     "dismissNotificationTitle": "收起此通知",
-    "editCommentTitle": "编辑自己发布的评论 (远端同步)",
     "fileCount_other": "{{count}} 个文件",
-    "gotIt": "知道了",
     "loadChangedFilesFailed": "拉取变更文件列表失败",
     "loadCommentsFailed": "拉取评论失败",
     "loadingContentHint": "从本地镜像读 git blob，大文件 / 二进制判定时可能略慢",
@@ -286,7 +278,6 @@
     "preparingSync": "准备同步",
     "readFileContentFailed": "读取文件内容失败",
     "readFileContentFailedNamed": "读取 {{path}} 内容失败",
-    "reply": "回复",
     "resizeFileListTitle": "拖动调整文件树宽度",
     "retry": "重试",
     "scopeAll": "全部变更",
